First tinkering with learnign why reshares do not get shared to other networks
I have done some investigation and found that the reason Pump.io reshares do not get sent to other networks, even when the original message was public, are as follows:

1) The To: list loaded from PyPump v0.5 is empty
2) THe CC: list is not, but only contains my followers.

Thus, under the current privacy model implicit in NavierStokes, this is not shared because things CC'd to only followers are not treated as public. This needs more thinking.